Moved PDF to right place
